+++ crypto/tls/test/tls_integration_go_1_21_test.go
@@ -1,6 +1,0 @@
-//go:build go1.21
-
-package test
-
-// The error message changed in Go 1.21.
-const badCertificateErrorMessage = "remote error: tls: certificate required"